requirements.txt
# requirements.txt
pandas
pyarrow
fastapi
uvicorn[standard]
confluent-kafka
aiokafka
scikit-learn
lightgbm
joblib
pydantic>=2.0
PyJWT>=2.8
python-dotenv>=1.0
sqlmodel>=0.0.8
psycopg2-binary




